+++ source/Erwin/level/scene.h
@@ -16,7 +16,7 @@ public:
 	static void select(erwin::EntityID entity);
 	static void drop_selection();
 
-	static void mark_for_removal(erwin::EntityID entity, uint64_t reflected_component);
+	static void mark_for_removal(erwin::EntityID entity, uint32_t reflected_component);
 	static void mark_for_removal(erwin::EntityID entity);
 
 	static void cleanup();
